+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/StatusListIssuer.kt
@@ -31,7 +31,21 @@ interface StatusListIssuer : StatusIssuer, StatusProvider {
      * [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus.Invalid].
      * Returns true if this credential has been revoked.
      */
-    fun revokeCredentialByIndex(timePeriod: Int, statusListIndex: ULong): Boolean
+    fun revokeCredentialByIndex(timePeriod: Int, statusListIndex: Long): Boolean
+
+    @Deprecated(
+        "Use a Long statusListIndex",
+        ReplaceWith("revokeCredentialByIndex(timePeriod, statusListIndex.toLong())"),
+    )
+    fun revokeCredentialByIndex(timePeriod: Int, statusListIndex: ULong): Boolean =
+        revokeCredentialByIndex(
+            timePeriod = timePeriod,
+            statusListIndex = statusListIndex.toLong().also {
+                require(statusListIndex <= Long.MAX_VALUE.toULong()) {
+                    "statusListIndex must be at most Long.MAX_VALUE"
+                }
+            },
+        )
 
     /**
      * Sets the status of one specific credential to